% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_citbi.R
\name{all_citbi}
\alias{all_citbi}
\title{Always predict ciTBI}
\usage{
all_citbi(data)
}
\arguments{
\item{data}{A data frame with one row per patient}
}
\value{
A factor vector of predictions ("ciTBI")
}
\description{
Always predict ciTBI
}
