% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_ct_data.R
\name{clean_ct_data}
\alias{clean_ct_data}
\title{Clean CT data}
\usage{
clean_ct_data(df)
}
\arguments{
\item{df}{A data frame with ciTBI and predictors}
}
\value{
A cleaned tibble
}
\description{
This function cleans the citbi.csv data
}
