% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/no_citbi.R
\name{no_citbi}
\alias{no_citbi}
\title{Always predict no ciTBI}
\usage{
no_citbi(data)
}
\arguments{
\item{data}{A data frame with one row per patient}
}
\value{
A factor vector of predictions ("no_ciTBI")
}
\description{
Always predict no ciTBI
}
